// packages/prisma-thinkthank/generated/zod/inputTypeSchemas/ProjectInviteMinOrderByAggregateInputSchema.ts
import type { Prisma } from '@prisma/client';

import { z } from 'zod';
import { SortOrderSchema } from './SortOrderSchema';

export const ProjectInviteMinOrderByAggregateInputSchema: z.ZodType<Prisma.ProjectInviteMinOrderByAggregateInput> = z.object({
  email: z.lazy(() => SortOrderSchema).optional(),
  expires: z.lazy(() => SortOrderSchema).optional(),
  projectId: z.lazy(() => SortOrderSchema).optional(),
  role: z.lazy(() => SortOrderSchema).optional(),
  createdAt: z.lazy(() => SortOrderSchema).optional()
}).strict();

export default ProjectInviteMinOrderByAggregateInputSchema;